service and controller which were Further refine

+++ src/main/java/com/ye/vio/service/impl/RentServiceImpl.java
@@ -4,6 +4,8 @@ import com.ye.vio.dao.RentDao;
 
 import com.ye.vio.dto.RentCondition;
 import com.ye.vio.entity.Rent;
+import com.ye.vio.enums.CustomizeErrorCode;
+import com.ye.vio.exception.CustomizeException;
 import com.ye.vio.service.RentService;
 import com.ye.vio.util.PageUtil;
 import com.ye.vio.util.UUIDUtils;
@@ -37,6 +39,9 @@ public class RentServiceImpl implements RentService {
 
     @Override
     public List<Rent> getRentList(RentCondition rentCondition, int pageIndex, int pageSize) {
+        if (pageIndex<=0)
+            throw new CustomizeException(CustomizeErrorCode.INVALID_INPUT);
+
         int rowIndex= PageUtil.pageIndexToRowIndex(pageIndex,pageSize);
 
         return rentDao.queryRentList(rentCondition,rowIndex,pageSize);
@@ -44,6 +49,9 @@ public class RentServiceImpl implements RentService {
 
     @Override
     public List<Rent> getRentListByUserId(String userId, int pageIndex, int pageSize) {
+        if (userId==null||pageIndex<=0)
+            throw new CustomizeException(CustomizeErrorCode.INVALID_INPUT);
+
         int rowIndex= PageUtil.pageIndexToRowIndex(pageIndex,pageSize);
         return rentDao.queryRentListByUserId(userId,rowIndex,pageSize);
     }
@@ -51,17 +59,25 @@ public class RentServiceImpl implements RentService {
     @Override
     @Transactional
     public int addRent(Rent rent) {
+        if (rent.getUser().getUserId()==null)
+            throw new CustomizeException(CustomizeErrorCode.INVALID_INPUT);
+
         rent.setRentId(UUIDUtils.UUID());
-        UserVo userVo=new UserVo();
-        userVo.setUserId("1");
-        rent.setUser(userVo);
         rent.setCreateTime(new Date());
-        return rentDao.insertRent(rent);
+        int effected=rentDao.insertRent(rent);
+        if(effected<=0)throw new CustomizeException(CustomizeErrorCode.RENT_ADD_ERROR);
+
+        return effected;
     }
 
     @Override
     @Transactional
     public int removeRentByRentId(String rentId, String userId) {
-        return rentDao.deleteRentByRentId(rentId,userId);
+        if (rentId==null||userId==null)
+            throw new CustomizeException(CustomizeErrorCode.INVALID_INPUT);
+        int effected=rentDao.deleteRentByRentId(rentId,userId);
+        if(effected<=0)throw new CustomizeException(CustomizeErrorCode.RENT_DELETE_ERROR);
+
+        return effected;
     }
 }
